#5 #3 rearranging some of the structure of the app. Able to authenticate with firebase from Fluttergit add --all
