+++ front-end/cantrain/lib/main.dart
@@ -1,8 +1,9 @@
+import 'package:cantrain/views/main_page.dart';
 import 'package:flutter/material.dart';
-import '/views/login.dart';
+import 'views/login_page.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'firebase_options.dart';
-
+ 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
 
@@ -21,14 +22,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: Navigator(
-        pages: [
-          MaterialPage(child: LoginView())
-        ],
-        onPopPage: (route, result) {
-          return route.didPop(result);
-        },
-      ),
+      home: MainView(),
       title: 'CanTrain',
       theme: ThemeData(
         // This is the theme of your application.
